lambda-function/requirements.txt
pillow==11.3.0
boto3==1.40.6
